mixpanel actions: refactor the event properties
