+++ packages/destination-actions/src/destinations/mixpanel/trackEvent/functions.ts
@@ -1,9 +1,12 @@
 
+import { omit } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import dayjs from '../../../lib/dayjs'
 import { MixpanelEventProperties } from '../mixpanel-types'
 import { getBrowser, getBrowserVersion, cheapGuid } from '../utils'
+
+const mixpanelReservedProperties = ['time', 'id', '$anon_id', 'distinct_id', '$group_id', '$insert_id', '$user_id']
 
 export function getEventProperties(payload: Payload, settings: Settings): MixpanelEventProperties {
     const datetime = payload.time
@@ -19,6 +22,7 @@ export function getEventProperties(payload: Payload, settings: Settings): Mixpan
         time: time,
         ip: payload.ip,
         id: payload.distinct_id,
+        $anon_id: payload.anonymous_id,
         distinct_id: payload.distinct_id,
         $app_build_number: payload.app_build,
         $app_version_string: payload.app_version,
@@ -35,6 +39,7 @@ export function getEventProperties(payload: Payload, settings: Settings): Mixpan
         $device_type: payload.device_type,
         $device_name: payload.device_name,
         $group_id: payload.group_id,
+        $identified_id: payload.user_id,
         $insert_id: payload.insert_id ?? cheapGuid(),
         $ios_ifa: payload.idfa,
         $lib_version: payload.library_version,
@@ -48,6 +53,7 @@ export function getEventProperties(payload: Payload, settings: Settings): Mixpan
         $screen_width: payload.screen_width,
         $screen_density: payload.screen_density,
         $source: 'segment',
+        $user_id: payload.user_id,
         $wifi_enabled: payload.wifi,
         mp_country_code: payload.country,
         mp_lib: payload.library_name && `Segment: ${ payload.library_name }`,
@@ -57,6 +63,7 @@ export function getEventProperties(payload: Payload, settings: Settings): Mixpan
         utm_medium: utm.utm_medium,
         utm_source: utm.utm_source,
         utm_term: utm.utm_term,
-        ...payload.event_properties
+        // Ignore Mixpanel reserved properties
+        ...omit(payload.event_properties, mixpanelReservedProperties)
     }
 }
